Initial commit.
This is the barest JavaFX Framework, simply puts up a blank stage. However it doesn't have the memory access violation crash anymore which was a big deal to get rid of. I had to package the whole thing into a JAR file to get it to work. It could have to do with module or class path perhaps.